% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.alpha.R
\name{add.alpha}
\alias{add.alpha}
\title{Add transparency to a color}
\usage{
add.alpha(col, alpha = 1)
}
\arguments{
\item{col}{a vector of colors}

\item{alpha}{the level of transparency}
}
\value{
the vector of transparent colors
}
\description{
Add transparency to a color
}
