% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bary_Projector.R
\name{Bary_Projector}
\alias{Bary_Projector}
\title{Compute Projector matrix from DESIGN_rows}
\usage{
Bary_Projector(DESIGN_rows)
}
\arguments{
\item{DESIGN_rows}{Design matrix for the rows}
}
\value{
The projector matrix for the barycentric space
}
\description{
Compute Projector matrix from DESIGN_rows
}
