% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bary_Projector_Cond.R
\name{Bary_Projector_Cond}
\alias{Bary_Projector_Cond}
\title{Compute condenses Projector matrix from DESIGN_rows}
\usage{
Bary_Projector_Cond(DESIGN_rows)
}
\arguments{
\item{DESIGN_rows}{Design matrix for the rows}
}
\value{
The condensed projector matrix for the barycentric space
}
\description{
Compute condenses Projector matrix from DESIGN_rows
}
