% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bary_Projector_Cond_Sqrt.R
\name{Bary_Projector_Cond_Sqrt}
\alias{Bary_Projector_Cond_Sqrt}
\title{Compute condensed Projector matrix from DESIGN_rows}
\usage{
Bary_Projector_Cond_Sqrt(DESIGN_rows_mat)
}
\arguments{
\item{DESIGN_rows}{Design matrix for the rows}
}
\value{
The condensed projector matrix for the barycentric space
}
\description{
Compute condensed Projector matrix from DESIGN_rows
}
