% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSplus.R
\name{ComputeSplus}
\alias{ComputeSplus}
\title{Compute a compromise, as defined in DiSTATIS}
\usage{
ComputeSplus(CubeCP, alpha)
}
\arguments{
\item{CubeCP}{Array of (double-centered) cross-product matrices}

\item{alpha}{Vector of alpha-weights}
}
\value{
The compromise
}
\description{
Compute a compromise, as defined in DiSTATIS
}
