% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Condense_Rows.R
\name{Condense_Rows}
\alias{Condense_Rows}
\title{Condence the rows of a matrix according to DESIGN$rows}
\usage{
Condense_Rows(X, DESIGN_rows_mat)
}
\arguments{
\item{X}{the data matrix}

\item{DESIGN_rows_mat}{Column vector(s) to discriminate rows (nominal or colors or DESIGN)}
}
\value{
The row-averaged version of X
}
\description{
Condence the rows of a matrix according to DESIGN$rows
}
