% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP2MFAnormedCP.R
\name{CP2MFAnormedCP}
\alias{CP2MFAnormedCP}
\title{MFA normalize an array of cross-product matrices}
\usage{
CP2MFAnormedCP(CP3)
}
\arguments{
\item{CP3}{An array of cross-product matrices}
}
\value{
An array of MFA-normalized cross-product matrices
}
\description{
MFA normalize an array of cross-product matrices
}
