% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DblCenterD2.R
\name{DblCenterD2}
\alias{DblCenterD2}
\title{Double Center a squared distance matrix}
\usage{
DblCenterD2(D2)
}
\arguments{
\item{D2}{a squared distance matrix}
}
\value{
a cross product matrix
}
\description{
Double Center a squared distance matrix
}
