% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dev2.R
\name{Dev2}
\alias{Dev2}
\title{Compute the squared distances between rows of two matrices}
\usage{
Dev2(from, to)
}
\arguments{
\item{from}{The first matrix}

\item{to}{The second matrix}
}
\value{
The squared distances
}
\description{
Compute the squared distances between rows of two matrices
}
