% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiDiSTATIS_boot_tables.R
\name{DiDiSTATIS_boot_tables}
\alias{DiDiSTATIS_boot_tables}
\title{Bootstrap resample rows within hypothesized groups for DiDiSTATIS}
\usage{
DiDiSTATIS_boot_tables(input = input,
  Hierarchy_of_tables = Hierarchy_of_tables, res_BaryGrand = res_BaryGrand,
  niter = 100)
}
\arguments{
\item{input}{Items input to DiDiSTATIS}

\item{Hierarchy_of_tables}{Results related to computing Barycentric Group and Grand Compromises}

\item{res_BaryGrand}{Results from decomposing the Barycentric Grand Compromise}

\item{niter}{Number of bootstrap samples}
}
\value{
Bootstrap output
}
\description{
Bootstrap resample rows within hypothesized groups for DiDiSTATIS
}
