% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiDiSTATIS_perm_omni_sort.R
\name{DiDiSTATIS_perm_omni_sort}
\alias{DiDiSTATIS_perm_omni_sort}
\title{Permutation test for the omnibus null hypothesis for DiDiSTATIS sorting data}
\usage{
DiDiSTATIS_perm_omni_sort(input = NULL, Hierarchy_of_tables = NULL,
  res_BaryGrand = NULL, niter = 100)
}
\arguments{
\item{input}{Items input to HiDiSTATIS}

\item{Hierarchy_of_tables}{Results related to computing Group and Grand Compromises}

\item{res_BaryGrand}{Results from decomposing the Barycentric Grand Compromise}

\item{niter}{Number of iterations (permutations)}
}
\value{
Permuted output
}
\description{
Permutation test for the omnibus null hypothesis for DiDiSTATIS sorting data
}
