% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiDiSTATIS_SH_rows.R
\name{DiDiSTATIS_SH_rows}
\alias{DiDiSTATIS_SH_rows}
\title{Leave One Out cross-validation for the rows for DiDiSTATIS}
\usage{
DiDiSTATIS_SH_rows(input, Hierarchy_of_tables, res_BaryGrand, niter = 100)
}
\arguments{
\item{input}{A list of things input to DiDiSTATIS}

\item{Hierarchy_of_tables}{Output of GetBaryGrandComp()}

\item{res_BaryGrand}{Output of EigenDiDiSTATIS()}

\item{niter}{Number of iterations}
}
\value{
A list of the SH results
}
\description{
Leave One Out cross-validation for the rows for DiDiSTATIS
}
