% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiDiSTATIS_Summary_Stats_Table.R
\name{DiDiSTATIS_Summary_Stats_Table}
\alias{DiDiSTATIS_Summary_Stats_Table}
\title{Print a table of summary stats}
\usage{
DiDiSTATIS_Summary_Stats_Table(res_DiDiSTATIS, main = NULL)
}
\arguments{
\item{res_DiDiSTATIS}{DiDiSTATIS output with all inference}

\item{main}{Table title}
}
\value{
A table of summary stats
}
\description{
Print a table of summary stats
}
