% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiMDS.R
\name{DiMDS}
\alias{DiMDS}
\title{Conduct barycentric discriminant metric multidimensional scaling}
\usage{
DiMDS(DATA, data_are, DESIGN_rows, n2k = NULL, Perm_rows = F,
  Perm_rows_niter = 1000, LOO_rows = F, LOO_rows_multiplier = 10,
  SH_rows = F, SH_rows_niter = 50)
}
\arguments{
\item{DATA}{A distance-data matrix}

\item{data_are}{Flag to indicate data type #d, d2, CP, or X_to_cos, X_to_cov, X_to_cor}

\item{DESIGN_rows}{Column vector(s) to discriminate rows (nominal or colors or DESIGN)}

\item{n2k}{Number of components to keep}

\item{Perm_rows}{Boolean to conduct permutation test}

\item{Perm_rows_niter}{Number of permutation iterations}

\item{LOO_rows}{Boolean to conduct Leave One Out (LOO) cross-validation}

\item{LOO_rows_multiplier}{Increase the number of LOO iterations}

\item{SH_rows}{Boolean to conduct Split Half (SH) cross-validation}

\item{SH_rows_niter}{Number of SH iterations}
}
\value{
Results of DiMDS
}
\description{
Conduct barycentric discriminant metric multidimensional scaling
}
