% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiMDS_boot_rows.R
\name{DiMDS_boot_rows}
\alias{DiMDS_boot_rows}
\title{Bootstrap resample rows within hypothesized groups for DiMDS}
\usage{
DiMDS_boot_rows(input = NULL, res_Disc_Full = res_Disc_Full, niter = 10)
}
\arguments{
\item{input}{Items input to DiMDS}

\item{res_Disc_Full}{Output from decomposing the barycentric discriminant sub-space}

\item{niter}{Number of iterations (permutations)}
}
\value{
Permuted output
}
\description{
Bootstrap resample rows within hypothesized groups for DiMDS
}
