% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiMDS_LOO_rows.R
\name{DiMDS_LOO_rows}
\alias{DiMDS_LOO_rows}
\title{Leave One Out cross-validation for the rows for DiMDS}
\usage{
DiMDS_LOO_rows(input = NULL, res_Disc_Full = NULL, multiplier = 1)
}
\arguments{
\item{input}{Items input to DiMDS}

\item{res_Disc_Full}{Output from decomposing the barycentric discriminant sub-space}

\item{multiplier}{to increase the number of iterations}
}
\value{
Prediction_Random
}
\description{
Leave One Out cross-validation for the rows for DiMDS
}
