% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiMDS_perm_rows.R
\name{DiMDS_perm_rows}
\alias{DiMDS_perm_rows}
\title{Permutation test for the null hypothesis on the rows of DiMDS}
\usage{
DiMDS_perm_rows(input = NULL, res_Disc_Full = NULL, niter = 100)
}
\arguments{
\item{input}{Items input to DiMDS}

\item{res_Disc_Full}{Output from decomposing the barycentric discriminant sub-space}

\item{niter}{Number of iterations (permutations)}
}
\value{
Permuted output
}
\description{
Permutation test for the null hypothesis on the rows of DiMDS
}
