% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist2CP.R
\name{Dist2CP}
\alias{Dist2CP}
\title{Apply DblCenterDist to an array of squared distance matrices}
\usage{
Dist2CP(D3)
}
\arguments{
\item{D3}{An array of squared distance matrices}
}
\value{
An array of CP matrices
}
\description{
Apply DblCenterDist to an array of squared distance matrices
}
