% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFromSort.R
\name{DistanceFromSort}
\alias{DistanceFromSort}
\title{Transform Sorting data to a Distance Matrix}
\usage{
DistanceFromSort(X)
}
\arguments{
\item{X}{A rectangular matrix of sorting data (nominal; stimuli by participants)}
}
\value{
An array of (sorting) distance matrices
}
\description{
Transform Sorting data to a Distance Matrix
}
