% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSTATIS.R
\name{DiSTATIS}
\alias{DiSTATIS}
\title{Conduct DiSTATIS}
\usage{
DiSTATIS(D2_array, DESIGN_tables, n2k = NULL, main = NULL)
}
\arguments{
\item{D2_array}{An array of squared distance matrices}

\item{DESIGN_tables}{DESIGN matrix for the tables}

\item{n2k}{Number (of components) to keep}

\item{main}{Title for Factor Maps}
}
\value{
Factor maps, and a list of computational results
}
\description{
Conduct DiSTATIS
}
