% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenCP.R
\name{EigenCP}
\alias{EigenCP}
\title{Eigen-decompose a CP matrix.}
\usage{
EigenCP(CP, n2k = NULL)
}
\arguments{
\item{CP}{A cross-product matrix}

\item{n2k}{Number of components 2 keep}
}
\value{
Basic output and a PrettyPlot
}
\description{
Eigen-decompose a CP matrix.
}
