% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenCP_Full.R
\name{EigenCP_Full}
\alias{EigenCP_Full}
\title{Eigen-decompose a CP matrix.}
\usage{
EigenCP_Full(CP, DESIGN_rows = DESIGN_rows, n2k = NULL)
}
\arguments{
\item{CP}{A cross-product matrix}

\item{DESIGN_rows}{List of DESIGN info for the rows}

\item{n2k}{Number of components 2 keep}
}
\value{
Basic output and a PrettyPlot
}
\description{
Eigen-decompose a CP matrix.
}
