% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenDiCP.R
\name{EigenDiCP}
\alias{EigenDiCP}
\title{Discriminant Eigen-decomposition for a CP matrix.}
\usage{
EigenDiCP(CP, DESIGN_rows = NULL, n2k = NULL)
}
\arguments{
\item{CP}{A cross-product matrix}

\item{DESIGN_rows}{DESIGN matrix for the rows}

\item{n2k}{Number of components 2 keep}
}
\value{
Computational output and PrettyPlots
}
\description{
Discriminant Eigen-decomposition for a CP matrix.
}
