% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenDiDiSTATIS.R
\name{EigenDiDiSTATIS}
\alias{EigenDiDiSTATIS}
\title{Eigen-decompose the Barycentric Grand Compromise, and project in Groups}
\usage{
EigenDiDiSTATIS(Hierarchy_of_tables, DESIGN_rows, DESIGN_tables, n2k = NULL)
}
\arguments{
\item{Hierarchy_of_tables}{All output related to individuals, Groups, and Grand Compromise}

\item{DESIGN_rows}{The file structure related to the row DESIGN}

\item{DESIGN_tables}{The file structure related to the table DESIGN}

\item{n2k}{Number of components 2 keep}
}
\value{
Output
}
\description{
Eigen-decompose the Barycentric Grand Compromise, and project in Groups
}
