% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenDiSTATIS.R
\name{EigenDiSTATIS}
\alias{EigenDiSTATIS}
\title{Eigen-decompose the Compromise, and then compute everything else}
\usage{
EigenDiSTATIS(DiSTATIS_collapsed, DESIGN = NULL, n2k = NULL)
}
\arguments{
\item{DiSTATIS_collapsed}{All output related to individuals and Compromise}

\item{DESIGN}{The file structure related to the (table, and if applicable, row) DESIGN}

\item{n2k}{Number of components 2 keep}
}
\value{
Computational output
}
\description{
Eigen-decompose the Compromise, and then compute everything else
}
