% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenHiDiSTATIS.R
\name{EigenHiDiSTATIS}
\alias{EigenHiDiSTATIS}
\title{Eigen-decompose the Grand Compromise, and then compute everything else}
\usage{
EigenHiDiSTATIS(HiDiSTATIS_collapsed, DESIGN_tables = NULL, n2k = NULL)
}
\arguments{
\item{HiDiSTATIS_collapsed}{All output related to individuals, Groups, and Grand Compromise}

\item{DESIGN_tables}{The file structure related to the table DESIGN}

\item{n2k}{Number of components 2 keep}
}
\value{
Computational output
}
\description{
Eigen-decompose the Grand Compromise, and then compute everything else
}
