% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenHMFA.R
\name{EigenHMFA}
\alias{EigenHMFA}
\title{Eigen-decompose the Grand Consensus, and then compute everything else}
\usage{
EigenHMFA(HMFA_collapsed, DESIGN_rows, DESIGN_tables, n2k = NULL)
}
\arguments{
\item{HMFA_collapsed}{All output related to individuals and Consensus}

\item{DESIGN_rows}{List of DESIGN info for rows}

\item{DESIGN_tables}{List of DESIGN info for tables}

\item{n2k}{Number of components 2 keep}
}
\value{
Computational output
}
\description{
Eigen-decompose the Grand Consensus, and then compute everything else
}
