% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlipCheck.R
\name{FlipCheck}
\alias{FlipCheck}
\title{If Component 1 factor Scores are all on the left side, flip them to the right side}
\usage{
FlipCheck(F)
}
\arguments{
\item{F}{Factor scores}
}
\value{
Flip-checked factor scores
}
\description{
If Component 1 factor Scores are all on the left side, flip them to the right side
}
