% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAlpha.R
\name{GetAlpha}
\alias{GetAlpha}
\title{Obtain a vector of alpha-weights from an inner product matrix (RV or C)}
\usage{
GetAlpha(C)
}
\arguments{
\item{C}{The inner product (RV or C) matrix}
}
\value{
a vector of alpha-weights
}
\description{
Obtain a vector of alpha-weights from an inner product matrix (RV or C)
}
