% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBaryGrandComp.R
\name{GetBaryGrandComp}
\alias{GetBaryGrandComp}
\title{Integrate tables hierarchically, with respect to row design,
to give barycentric group and grand compormises}
\usage{
GetBaryGrandComp(CP_array, DESIGN_rows, DESIGN_tables, MFA1_Flag = TRUE,
  RV1_Flag = TRUE, MFA2_Flag = TRUE, RV2_Flag = TRUE)
}
\arguments{
\item{CP_array}{An array of CP matrices}

\item{DESIGN_rows}{List of DESIGN info for rows}

\item{DESIGN_tables}{List of DESIGN info for tables}

\item{MFA1_Flag}{#TRUE gives MFA-norm; FALSE gives no norm}

\item{RV1_Flag}{# TRUE gives RV; FALSE gives C; "skip" sets alpha-weights to 1/CD (1/N)}

\item{MFA2_Flag}{#TRUE gives MFA-norm; FALSE gives no norm}

\item{RV2_Flag}{# TRUE gives RV; FALSE gives C; "skip" sets alpha-weights to 1/D}
}
\value{
A list of compromises and other computed objects
}
\description{
Integrate tables hierarchically, with respect to row design,
to give barycentric group and grand compormises
}
