% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCmat.R
\name{GetCmat}
\alias{GetCmat}
\title{Compute an inner-product matrix (RV or C) from an array of CP matrices}
\usage{
GetCmat(CubeCP, RV = FALSE)
}
\arguments{
\item{CubeCP}{an array of CP matrices}

\item{RV}{flag to normalize. TRUE gives RV. FALSE gives standard inner product, C.}
}
\value{
Inner product matrix
}
\description{
Compute an inner-product matrix (RV or C) from an array of CP matrices
}
