% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCompromise.R
\name{GetCompromise}
\alias{GetCompromise}
\title{Collapse tables for DiSTATIS}
\usage{
GetCompromise(CP_array, DESIGN_tables)
}
\arguments{
\item{DESIGN_tables}{Column vector(s) to color tables (nominal or colors or DESIGN)}

\item{D2_array}{A array of squared distance matrices}

\item{data_are}{Flag to indicate data type #d, d2, CP, or X_to_cos, X_to_cov, X_to_cor}

\item{n2k}{Number of components to keep}

\item{main}{Title for Factor Maps}
}
\value{
Factor maps, and a list of computational results
}
\description{
Collapse tables for DiSTATIS
}
