% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCP.R
\name{GetCP}
\alias{GetCP}
\title{Transform from input data to a CP matrix}
\usage{
GetCP(DATA, data_are = NULL)
}
\arguments{
\item{DATA}{The data}

\item{data_are}{Flag to indicate data type #d, d2, CP, or X_to_cos, X_to_cov, X_to_cor}
}
\value{
Cross-product matrix
}
\description{
Transform from input data to a CP matrix
}
