% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCP_array.R
\name{GetCP_array}
\alias{GetCP_array}
\title{Transform from input data to a CP_array}
\usage{
GetCP_array(DATA, data_are = NULL)
}
\arguments{
\item{DATA}{The data}

\item{data_are}{Flag to indicate data type #sort, sort_dist, d_array, d2_array, CP_array}
}
\value{
An array of cross-product matrices
}
\description{
Transform from input data to a CP_array
}
