% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGrandCompromise.R
\name{GetGrandCompromise}
\alias{GetGrandCompromise}
\title{Integrate tables hierarchically, to give group and grand compormises}
\usage{
GetGrandCompromise(CP_array, DESIGN_tables)
}
\arguments{
\item{CP_array}{An array of CP matrices}

\item{DESIGN_tables}{List of DESIGN info for tables}
}
\value{
A list of compromises and other computed objects
}
\description{
Integrate tables hierarchically, to give group and grand compormises
}
