% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGrandCompromise_BootTables.R
\name{GetGrandCompromise_BootTables}
\alias{GetGrandCompromise_BootTables}
\title{Integrate tables hierarchically, to give group and grand compormises
Differs from GetGrandCompromise.R by computing alpha2_supp,
because bootstrapped group compormises are projected as supplementary
and so need to be weighted as supplementary to the original Rv_D space.}
\usage{
GetGrandCompromise_BootTables(CP_array, DESIGN_tables, Hierarchy_of_tables)
}
\arguments{
\item{CP_array}{An array of CP matrices}

\item{DESIGN_tables}{List of DESIGN info for tables}

\item{Hierarchy_of_tables}{Results related to computing Group and Grand Compromises}
}
\value{
A list of compromises and other computed objects
}
\description{
Integrate tables hierarchically, to give group and grand compormises
Differs from GetGrandCompromise.R by computing alpha2_supp,
because bootstrapped group compormises are projected as supplementary
and so need to be weighted as supplementary to the original Rv_D space.
}
