% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Happy_p.R
\name{Happy_p}
\alias{Happy_p}
\title{Remove the leading zero before the p-value}
\usage{
Happy_p(Distribution, Stat)
}
\arguments{
\item{Distribution}{The distribution of values}

\item{Stat}{The statistic to compare to the distribution}
}
\value{
APA format
}
\description{
Remove the leading zero before the p-value
}
