% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDiSTATIS.R
\name{HiDiSTATIS}
\alias{HiDiSTATIS}
\title{Conduct Hierarchical DiSTATIS}
\usage{
HiDiSTATIS(DATA, data_are = NULL, DESIGN_rows, DESIGN_tables, n2k = NULL,
  Perm_tables = F, Perm_tables_niter = 1000, Boot_tables = F,
  Boot_tables_niter = 1000, LOO_tables = F, LOO_tables_multiplier = 10,
  SH_tables = F, SH_tables_niter = 50)
}
\arguments{
\item{DATA}{The data}

\item{data_are}{Flag to indicate data type #sort, d_array, d2_array, CP_array}

\item{DESIGN_rows}{DESIGN info for the rows}

\item{DESIGN_tables}{DESIGN info for the tables}

\item{n2k}{Number (of components) to keep}

\item{Perm_tables}{Boolean to conduct permutation test}

\item{Perm_tables_niter}{Number of permutation iterations}

\item{Boot_tables}{Boolean to conduct bootstrap resampling}

\item{Boot_tables_niter}{Number of bootstrap iterations}

\item{LOO_tables}{Boolean to conduct Leave One Out (LOO) cross-validation}

\item{LOO_tables_multiplier}{Increase the number of LOO iterations}

\item{SH_tables}{Boolean to conduct Split Half (SH) cross-validation}

\item{SH_tables_niter}{Number of SH iterations}
}
\value{
Factor maps, and a list of computational results
}
\description{
Conduct Hierarchical DiSTATIS
}
