% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDiSTATIS_boot_tables.R
\name{HiDiSTATIS_boot_tables}
\alias{HiDiSTATIS_boot_tables}
\title{Bootstrap resample rows within hypothesized groups for HiDiSTATIS}
\usage{
HiDiSTATIS_boot_tables(input = input,
  Hierarchy_of_tables = Hierarchy_of_tables, res_GrandComp = res_GrandComp,
  niter = 10)
}
\arguments{
\item{input}{Items input to HiDiSTATIS}

\item{Hierarchy_of_tables}{Results related to computing Group and Grand Compromises}

\item{res_GrandComp}{Results from decomposing the Grand Compromise}

\item{niter}{Number of bootstrap samples}
}
\value{
Bootstrap output
}
\description{
Bootstrap resample rows within hypothesized groups for HiDiSTATIS
}
