% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDiSTATIS_LOO_tables.R
\name{HiDiSTATIS_LOO_tables}
\alias{HiDiSTATIS_LOO_tables}
\title{Leave One Out cross-validation for the tables of HiDiSTATIS}
\usage{
HiDiSTATIS_LOO_tables(input = NULL, Hierarchy_of_tables = NULL,
  res_GrandComp = NULL, LOO_tables_multiplier = 1)
}
\arguments{
\item{input}{Items input to HiDiSTATIS}

\item{Hierarchy_of_tables}{Results related to computing Group and Grand Compromises}

\item{res_GrandComp}{Results from decomposing the Grand Compromise}

\item{LOO_tables_multiplier}{to increase the number of iterations}
}
\value{
Prediction_Random
}
\description{
Leave One Out cross-validation for the tables of HiDiSTATIS
}
