% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDiSTATIS_perm_tables.R
\name{HiDiSTATIS_perm_tables}
\alias{HiDiSTATIS_perm_tables}
\title{Permutation test for the null hypothesis on the tables of HiDiSTATIS}
\usage{
HiDiSTATIS_perm_tables(input = NULL, Hierarchy_of_tables = NULL,
  res_GrandComp = NULL, niter = 100)
}
\arguments{
\item{input}{Items input to HiDiSTATIS}

\item{Hierarchy_of_tables}{Results related to computing Group and Grand Compromises}

\item{res_GrandComp}{Results from decomposing the Grand Compromise}

\item{niter}{Number of iterations (permutations)}
}
\value{
Permuted output
}
\description{
Permutation test for the null hypothesis on the tables of HiDiSTATIS
}
