% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiDiSTATIS_SH_tables.R
\name{HiDiSTATIS_SH_tables}
\alias{HiDiSTATIS_SH_tables}
\title{Split Half cross-validation for the tables of HiDiSTATIS}
\usage{
HiDiSTATIS_SH_tables(input = NULL, Hierarchy_of_tables = NULL,
  res_GrandComp = NULL, niter = 1)
}
\arguments{
\item{input}{Items input to HiDiSTATIS}

\item{Hierarchy_of_tables}{Results related to computing Group and Grand Compromises}

\item{res_GrandComp}{Results from decomposing the Grand Compromise}

\item{niter}{the number of iterations}
}
\value{
Prediction_Random
}
\description{
Split Half cross-validation for the tables of HiDiSTATIS
}
