% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize_DESIGNs.R
\name{Initialize_DESIGNs}
\alias{Initialize_DESIGNs}
\title{Create Row and Table DESIGN objects}
\usage{
Initialize_DESIGNs(DESIGN_rows_mat = NULL, DESIGN_tables_mat = NULL)
}
\arguments{
\item{DESIGN_rows_mat}{DESIGN matrix for rows (stimuli)}

\item{DESIGN_tables_mat}{DESIGN matrix for tables (participants)}
}
\description{
Create Row and Table DESIGN objects
}
