% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFlip_mat.R
\name{makeFlip_mat}
\alias{makeFlip_mat}
\title{Flip_mat is a matrix to manually flip components to match results across methods}
\usage{
makeFlip_mat(Flip_axis1, Flip_axis2)
}
\arguments{
\item{Flip_axis1}{Boolean to Flip Factor Scores of axis1}

\item{Flip_axis2}{Boolean to Flip Factor Scores of axis2}
}
\value{
Flip_mat
}
\description{
Flip_mat is a matrix to manually flip components to match results across methods
}
