% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS.R
\name{MDS}
\alias{MDS}
\title{Conduct metric multidimensional scaling}
\usage{
MDS(DATA, data_are = NULL, DESIGN_rows, n2k = NULL)
}
\arguments{
\item{DATA}{The data}

\item{data_are}{Flag to indicate data type #d, d2, CP}

\item{DESIGN_rows}{List of DESIGN info}

\item{n2k}{Number of components to keep}
}
\value{
Results of MDS
}
\description{
Conduct metric multidimensional scaling
}
