% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFA.R
\name{MFA}
\alias{MFA}
\title{Conduct MFA (specifically for squared distance matrices)}
\usage{
MFA(D2_array, DESIGN_tables, n2k = NULL, main = NULL)
}
\arguments{
\item{D2_array}{An array of squared distance matrices}

\item{DESIGN_tables}{DESIGN matrix for the tables}

\item{n2k}{Number (of components) to keep}

\item{main}{Title for Factor Maps}
}
\value{
Factor maps, and a list of computational results
}
\description{
Conduct MFA (specifically for squared distance matrices)
}
