% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFAnormCP.R
\name{MFAnormCP}
\alias{MFAnormCP}
\title{MFA normalize a psd matrix}
\usage{
MFAnormCP(psd)
}
\arguments{
\item{psd}{A psd (such as cross-product) matrix}
}
\value{
An MFA-normalized psd matrix
}
\description{
MFA normalize a psd matrix
}
