% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFAnormCPFinder.R
\name{MFAnormCPFinder}
\alias{MFAnormCPFinder}
\title{Identify coefficients to MFA normalize an array}
\usage{
MFAnormCPFinder(CP_array)
}
\arguments{
\item{CP_array}{An array of cross-product matrices}
}
\value{
a vector of MFAnorm coefficients
}
\description{
Identify coefficients to MFA normalize an array
}
