% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFAnormCPFinder_1.R
\name{MFAnormCPFinder_1}
\alias{MFAnormCPFinder_1}
\title{Identify the coefficient to MFA normalize a matrix}
\usage{
MFAnormCPFinder_1(Y)
}
\arguments{
\item{Y}{A cross-product matrix}
}
\value{
The first eigen-value of Y
}
\description{
Identify the coefficient to MFA normalize a matrix
}
