% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MusicFreeSort}
\alias{MusicFreeSort}
\title{Free Sort of 36 Musical Pieces}
\format{The Music Free Sorting Data
\describe{
  \item{AlltheItems}{Sorting}
}}
\usage{
MusicFreeSort
}
\description{
Data collected from the laboratory of Dr. W.J. Dowling
The University of Texas at Dallas
}
\keyword{datasets}
