% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ortho_Projector.R
\name{Ortho_Projector}
\alias{Ortho_Projector}
\title{Compute the Orthogonal Projector matrix from the Barycentric Projector Matrix}
\usage{
Ortho_Projector(Pb)
}
\arguments{
\item{Pb}{Barycentric Projector Matrix}
}
\value{
The projector matrix for the orthogonal space
}
\description{
Compute the Orthogonal Projector matrix from the Barycentric Projector Matrix
}
