% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pick_rows_to_leave_out.R
\name{Pick_rows_to_leave_out}
\alias{Pick_rows_to_leave_out}
\title{Identify a matrix of row indices for LOO_rows}
\usage{
Pick_rows_to_leave_out(DESIGN_rows = DESIGN_rows, multiplier = 1)
}
\arguments{
\item{DESIGN_rows}{DESIGN matrix for rows (stimuli nested in categories)}

\item{multiplier}{parameter to increase the number of iterations}
}
\value{
A matrix of row indices for each iteration of the LOO
}
\description{
Identify a matrix of row indices for LOO_rows
}
