% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pick_rows_to_split.R
\name{Pick_rows_to_split}
\alias{Pick_rows_to_split}
\title{Identify a matrix of row indices for SH_rows}
\usage{
Pick_rows_to_split(DESIGN_rows = DESIGN_rows, niter = niter)
}
\arguments{
\item{DESIGN_rows}{DESIGN matrix for rows (stimuli nested in categories)}

\item{niter}{Number of iterations}
}
\value{
A matrix of row indices for each iteration of the SH
}
\description{
Identify a matrix of row indices for SH_rows
}
