% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pick_tables_to_leave_out.R
\name{Pick_tables_to_leave_out}
\alias{Pick_tables_to_leave_out}
\title{Identify a matrix of indices for LOO_tables}
\usage{
Pick_tables_to_leave_out(DESIGN_tables = DESIGN_tables, multiplier = 1)
}
\arguments{
\item{DESIGN_tables}{DESIGN info for tables (participants nested in groups)}

\item{multiplier}{parameter to increase the number of iterations}
}
\value{
A matrix of table indices for each iteration of the LOO
}
\description{
Identify a matrix of indices for LOO_tables
}
