% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pick_tables_to_split.R
\name{Pick_tables_to_split}
\alias{Pick_tables_to_split}
\title{Identify a matrix of indices for SH_talbes
The rows of the matrix produced correspond to the tables to leave out in each iteration}
\usage{
Pick_tables_to_split(DESIGN_tables = DESIGN_tables, niter = niter)
}
\arguments{
\item{DESIGN_tables}{DESIGN matrix for tables (participants nested in groups)}

\item{niter}{Number of iterations}
}
\value{
A matrix of tables indices to leave out for each iteration of the SH
}
\description{
Identify a matrix of indices for SH_talbes
The rows of the matrix produced correspond to the tables to leave out in each iteration
}
