% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotConfusion.R
\name{PlotConfusion}
\alias{PlotConfusion}
\title{Plot a confusion matrix}
\usage{
PlotConfusion(Confusion_mat = NULL)
}
\arguments{
\item{Confusion_mat}{Confusion matrix}
}
\value{
A pretty confusion matrix
}
\description{
Plot a confusion matrix
}
