% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiDiSTATIS.R
\name{PlotDiDiSTATIS}
\alias{PlotDiDiSTATIS}
\title{Plot results of DiDiSTATIS}
\usage{
PlotDiDiSTATIS(res_DiDiSTATIS, axes = c(1, 2), main = NULL)
}
\arguments{
\item{res_DiDiSTATIS}{The output of EigenHiDiSTATIS}

\item{axes}{Axes to plot. By default, c(1,2)}

\item{main}{Title of factor map}
}
\description{
Plot results of DiDiSTATIS
}
