% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiDiSTATIS_Boot_Tables.R
\name{PlotDiDiSTATIS_Boot_Tables}
\alias{PlotDiDiSTATIS_Boot_Tables}
\title{Plot results from bootstrap resampling of Tables for DiDiSTATIS}
\usage{
PlotDiDiSTATIS_Boot_Tables(res_DiDiSTATIS)
}
\arguments{
\item{res_DiDiSTATIS}{Output from DiDiSTATIS with Boot_tables==TRUE}
}
\value{
Output
}
\description{
Plot results from bootstrap resampling of Tables for DiDiSTATIS
}
