% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiDiSTATIS_perm_omni_sort.R
\name{PlotDiDiSTATIS_Perm_Omni}
\alias{PlotDiDiSTATIS_Perm_Omni}
\title{Plot results from omnibus test of Null Hypothesis}
\usage{
PlotDiDiSTATIS_Perm_Omni(res_DiDiSTATIS)
}
\arguments{
\item{res_DiDiSTATIS}{Output from DiDiSTATIS with Perm_omnibus==TRUE}
}
\value{
Histograms of the Dev2's
}
\description{
Plot results from omnibus test of Null Hypothesis
}
