% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiMDS.R
\name{PlotDiMDS}
\alias{PlotDiMDS}
\title{Plot results of DiMDS}
\usage{
PlotDiMDS(res_DiMDS, axes = c(1, 2))
}
\arguments{
\item{res_DiMDS}{The output of DiMDS}

\item{axes}{Axes to plot, default = c(1,2)}
}
\description{
Plot results of DiMDS
}
