% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiMDS_Boot_Rows.R
\name{PlotDiMDS_Boot_Rows}
\alias{PlotDiMDS_Boot_Rows}
\title{Plot results from Bootstrap on the Rows of DiMDS}
\usage{
PlotDiMDS_Boot_Rows(res_DiMDS)
}
\arguments{
\item{res_DiMDS}{Output from DiMDS}
}
\value{
95% Bootstrap confidence intervals
}
\description{
Plot results from Bootstrap on the Rows of DiMDS
}
