% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiMDS_LOO_Rows.R
\name{PlotDiMDS_LOO_Rows}
\alias{PlotDiMDS_LOO_Rows}
\title{Plot results from LOO on the Rows of DiMDS}
\usage{
PlotDiMDS_LOO_Rows(res_DiMDS)
}
\arguments{
\item{res_DiMDS}{Output from DiMDS}
}
\value{
LOO results
}
\description{
Plot results from LOO on the Rows of DiMDS
}
