% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiMDS_Perm_Rows.R
\name{PlotDiMDS_Perm_Rows}
\alias{PlotDiMDS_Perm_Rows}
\title{Plot results from test of Null Hypothesis on Rows}
\usage{
PlotDiMDS_Perm_Rows(res_DiMDS)
}
\arguments{
\item{res_DiMDS}{Output from DiMDS with Perm_rows==TRUE}
}
\value{
Histograms of the r2's
}
\description{
Plot results from test of Null Hypothesis on Rows
}
