% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiMDS_SH_Rows.R
\name{PlotDiMDS_SH_Rows}
\alias{PlotDiMDS_SH_Rows}
\title{Plot results from SH on the Rows of DiMDS}
\usage{
PlotDiMDS_SH_Rows(res_DiMDS)
}
\arguments{
\item{res_DiMDS}{Output from DiMDS}
}
\value{
SH results
}
\description{
Plot results from SH on the Rows of DiMDS
}
