% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiSTATIS.R
\name{PlotDiSTATIS}
\alias{PlotDiSTATIS}
\title{Plot results of DiSTATIS}
\usage{
PlotDiSTATIS(res_DiSTATIS, axes = c(1, 2), main = NULL)
}
\arguments{
\item{res_DiSTATIS}{The output of EigenDiSTATIS}

\item{axes}{Axes to plot. By default, c(1,2)}

\item{main}{Title of factor map}
}
\description{
Plot results of DiSTATIS
}
