% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHiDiSTATIS.R
\name{PlotHiDiSTATIS}
\alias{PlotHiDiSTATIS}
\title{Plot results of HiDiSTATIS}
\usage{
PlotHiDiSTATIS(res_HiDiSTATIS, axes = c(1, 2), main = NULL)
}
\arguments{
\item{res_HiDiSTATIS}{The output of EigenHiDiSTATIS}

\item{axes}{Axes to plot. By default, c(1,2)}

\item{main}{Title of factor map}
}
\description{
Plot results of HiDiSTATIS
}
