% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHiDiSTATIS_Boot_Tables.R
\name{PlotHiDiSTATIS_Boot_Tables}
\alias{PlotHiDiSTATIS_Boot_Tables}
\title{Plot results from bootstrap resampling of Tables for HiDiSTATIS}
\usage{
PlotHiDiSTATIS_Boot_Tables(res_HiDiSTATIS)
}
\arguments{
\item{res_HiDiSTATIS}{Output from HiDiSTATIS with Boot_tables==TRUE}
}
\value{
Output
}
\description{
Plot results from bootstrap resampling of Tables for HiDiSTATIS
}
