% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHiDiSTATIS_LOO_Tables.R
\name{PlotHiDiSTATIS_LOO_Tables}
\alias{PlotHiDiSTATIS_LOO_Tables}
\title{Plot results from LOO cross-validation of Tables for HiDiSTATIS}
\usage{
PlotHiDiSTATIS_LOO_Tables(res_HiDiSTATIS)
}
\arguments{
\item{res_HiDiSTATIS}{Output from HiDiSTATIS with LOO_tables==TRUE}
}
\value{
Output
}
\description{
Plot results from LOO cross-validation of Tables for HiDiSTATIS
}
