% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHiDiSTATIS_Perm_Tables.R
\name{PlotHiDiSTATIS_Perm_Tables}
\alias{PlotHiDiSTATIS_Perm_Tables}
\title{Plot results from test of Null Hypothesis on Tables}
\usage{
PlotHiDiSTATIS_Perm_Tables(res_HiDiSTATIS)
}
\arguments{
\item{res_HiDiSTATIS}{Output from HiDiSTATIS with Perm_tables==TRUE}
}
\value{
Histograms of the Dev2's
}
\description{
Plot results from test of Null Hypothesis on Tables
}
