% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHiDiSTATIS_SH_Tables.R
\name{PlotHiDiSTATIS_SH_Tables}
\alias{PlotHiDiSTATIS_SH_Tables}
\title{Plot results from SH cross-validation of Tables for HiDiSTATIS}
\usage{
PlotHiDiSTATIS_SH_Tables(res_HiDiSTATIS)
}
\arguments{
\item{res_HiDiSTATIS}{Output from HiDiSTATIS with SH_tables==TRUE}
}
\value{
Output
}
\description{
Plot results from SH cross-validation of Tables for HiDiSTATIS
}
