% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHMFA.R
\name{PlotHMFA}
\alias{PlotHMFA}
\title{Plot results of HMFA}
\usage{
PlotHMFA(res_HMFA, axes = c(1, 2), main = NULL)
}
\arguments{
\item{res_HMFA}{The output of EigenHMFA}

\item{axes}{Axes to plot. By default, c(1,2)}

\item{main}{Title of factor map}
}
\description{
Plot results of HMFA
}
