% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMFA.R
\name{PlotMFA}
\alias{PlotMFA}
\title{Plot results of MFA}
\usage{
PlotMFA(res_MFA, axes = c(1, 2), main = NULL)
}
\arguments{
\item{res_MFA}{The output of EigenMFA}

\item{axes}{Axes to plot. By default, c(1,2)}

\item{main}{Title of factor map}
}
\description{
Plot results of MFA
}
