% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_array_2_matrix.R
\name{rbind_array_2_matrix}
\alias{rbind_array_2_matrix}
\title{For permutation of tables of DiDiSTATIS, Plot each stimulus in the Grand Compromise}
\usage{
rbind_array_2_matrix(the_array)
}
\arguments{
\item{the_array}{a 3-d or 4-d data array}
}
\value{
the matrix
}
\description{
For permutation of tables of DiDiSTATIS, Plot each stimulus in the Grand Compromise
}
