% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segments_from_to.R
\name{Segments_from_to}
\alias{Segments_from_to}
\title{Shortcut to draw line segments between points}
\usage{
Segments_from_to(From, To, XandY = c(1, 2), col = "black", lwd = 1)
}
\arguments{
\item{From}{the first object}

\item{To}{the second object #"origin"}

\item{XandY}{the 2 axes to plot, concatenated}

\item{col}{line color}

\item{lwd}{line width}
}
\value{
Draw segments
}
\description{
Shortcut to draw line segments between points
}
