% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortDist2Sort.R
\name{SortDist2Sort}
\alias{SortDist2Sort}
\title{Transform a Sorting Distance matrix to a column of the Sorting matrix}
\usage{
SortDist2Sort(SortDist)
}
\arguments{
\item{SortDist}{A given sorting distance matrix}
}
\value{
A column of the sorting matrix
}
\description{
Transform a Sorting Distance matrix to a column of the Sorting matrix
}
