% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortFromSortDist.R
\name{SortFromSortDist}
\alias{SortFromSortDist}
\title{Transform from an array of Sorting Distance to a Sorting matrix}
\usage{
SortFromSortDist(SortDistArray)
}
\arguments{
\item{SortDistArray}{Array of sorting distance matrices}
}
\value{
The sorting matrix
}
\description{
Transform from an array of Sorting Distance to a Sorting matrix
}
