% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiDiSTATIS_LOO_rows.R
\name{DiDiSTATIS_LOO_rows}
\alias{DiDiSTATIS_LOO_rows}
\title{Leave One Out cross-validation for the rows for DiDiSTATIS}
\usage{
DiDiSTATIS_LOO_rows(input, Hierarchy_of_tables, res_BaryGrand, multiplier = 1)
}
\arguments{
\item{input}{A list of things input to DiDiSTATIS}

\item{Hierarchy_of_tables}{Output of GetBaryGrandComp()}

\item{res_BaryGrand}{Output of EigenDiDiSTATIS()}

\item{multiplier}{to increase the number of iterations}
}
\value{
A list of the LOO results
}
\description{
Leave One Out cross-validation for the rows for DiDiSTATIS
}
