% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiDiSTATIS_SH_rows.R
\name{DiDiSTATIS_SH_rows}
\alias{DiDiSTATIS_SH_rows}
\title{Leave One Out cross-validation for the rows for DiDiSTATIS}
\usage{
DiDiSTATIS_SH_rows(input, Hierarchy_of_tables, res_BaryGrand, niter = 100)
}
\arguments{
\item{input}{Items input to DiDiSTATIS}

\item{niter}{Number of iterations}
}
\value{
Prediction_Random
}
\description{
Leave One Out cross-validation for the rows for DiDiSTATIS
}
