% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiMDS_SH_rows.R
\name{DiMDS_SH_rows}
\alias{DiMDS_SH_rows}
\title{Split Half cross-validation for the rows for DiMDS}
\usage{
DiMDS_SH_rows(input = NULL, res_Disc_Full = NULL, niter = 1)
}
\arguments{
\item{input}{Items input to DiMDS}

\item{res_Disc_Full}{Output from decomposing the barycentric discriminant sub-space}

\item{niter}{The number of iterations}
}
\value{
Prediction_Random
}
\description{
Split Half cross-validation for the rows for DiMDS
}
