% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiSTATIS.R
\name{DiSTATIS}
\alias{DiSTATIS}
\title{Conduct DiSTATIS}
\usage{
DiSTATIS(DATA, data_are, DESIGN_rows, DESIGN_tables, n2k = NULL)
}
\arguments{
\item{DATA}{The data}

\item{data_are}{Flag to indicate data type #sort, d_array, d2_array, CP_array}

\item{DESIGN_rows}{List of DESIGN info for rows}

\item{DESIGN_tables}{List of DESIGN info for tables}

\item{n2k}{Number (of components) to keep}
}
\value{
A list of computational results
}
\description{
Conduct DiSTATIS
}
