% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenDiSTATIS.R
\name{EigenDiSTATIS}
\alias{EigenDiSTATIS}
\title{Eigen-decompose the Compromise, and then compute everything else}
\usage{
EigenDiSTATIS(DiSTATIS_collapsed, DESIGN_rows = NULL, DESIGN_tables,
  n2k = NULL)
}
\arguments{
\item{DiSTATIS_collapsed}{All output related to individuals and Consensus}

\item{DESIGN_rows}{List of DESIGN info for rows}

\item{DESIGN_tables}{List of DESIGN info for tables}

\item{n2k}{Number of components 2 keep}
}
\value{
Computational output
}
\description{
Eigen-decompose the Compromise, and then compute everything else
}
