% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenHMFA.R
\name{EigenHMFA}
\alias{EigenHMFA}
\title{Eigen-decompose the Grand Consensus, and then compute everything else}
\usage{
EigenHMFA(HMFA_collapsed, DESIGN = NULL, n2k = NULL)
}
\arguments{
\item{HMFA_collapsed}{All output related to individuals, Groups, and Grand Consensus}

\item{DESIGN}{The file structure related to the (table, and if applicable, row) DESIGN}

\item{n2k}{Number of components 2 keep}
}
\value{
Computational output
}
\description{
Eigen-decompose the Grand Consensus, and then compute everything else
}
