% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenMFA.R
\name{EigenMFA}
\alias{EigenMFA}
\title{Eigen-decompose the Consensus, and then compute everything else}
\usage{
EigenMFA(MFA_collapsed, DESIGN = NULL, n2k = NULL)
}
\arguments{
\item{MFA_collapsed}{All output related to individuals and Consensus}

\item{DESIGN}{The file structure related to the (table, and if applicable, row) DESIGN}

\item{n2k}{Number of components 2 keep}
}
\value{
Computational output
}
\description{
Eigen-decompose the Consensus, and then compute everything else
}
