% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGrandConsensus.R
\name{GetGrandConsensus}
\alias{GetGrandConsensus}
\title{Collapse tables for HMFA}
\usage{
GetGrandConsensus(CP_array, DESIGN_rows, DESIGN_tables)
}
\arguments{
\item{CP_array}{A array of cross-product matrices}

\item{DESIGN_rows}{List of DESIGN info for rows}

\item{DESIGN_tables}{List of DESIGN info for tables}
}
\value{
A list of compromises and other computed objects
}
\description{
Collapse tables for HMFA
}
