% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS.R
\name{MDS}
\alias{MDS}
\title{Conduct metric multidimensional scaling}
\usage{
MDS(D, data_are = NULL, DESIGN, n2k = NULL, main = NULL)
}
\arguments{
\item{D}{Traditionally, a distance matrix}

\item{data_are}{Flag to indicate data type #d, d2, CP}

\item{DESIGN}{List of DESIGN info}

\item{n2k}{Number of components to keep}

\item{main}{Title for Factor Maps}
}
\value{
Factor maps, and a list of computational results
}
\description{
Conduct metric multidimensional scaling
}
